moved 'dump_path_to_obj' func to 'json_helpers.py'
